# specs/003-user-auth/contracts/auth.openapi.yaml
openapi: 3.1.0
info:
  title: FixFlow Authentication API
  version: 2.0.0
  description: Account-request registration, administrator approval, and authenticated session contract.
servers:
  - url: /api
paths:
...
  /admin/account-requests/{account}/approve:
    patch:
      operationId: approveAccountRequest
      summary: Approve one pending public-role account request
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Account approved with immutable reviewer audit fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountRequestSuccess' }
        '401': { $ref: '#/components/responses/UnauthenticatedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /admin/account-requests/{account}/reject:
    patch:
      operationId: rejectAccountRequest
      summary: Reject one pending public-role account request
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                rejection_reason:
                  type: [string, 'null']
                  maxLength: 1000
      responses:
        '200':
          description: Account rejected with immutable reviewer audit fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountRequestSuccess' }
        '401': { $ref: '#/components/responses/UnauthenticatedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
        '409': { $ref: '#/components/responses/ConflictError' }
        '422': { $ref: '#/components/responses/ValidationError' }
components:
  parameters:
    AccountId:
      in: path
      name: account
      required: true
      schema: { type: integer, minimum: 1 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Reviewer:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
    AccountRequest:
      type: object
      additionalProperties: false
      required: [id, name, email, requested_role, status, registered_at, approved_by, approved_at, rejected_by, rejected_at, rejection_reason]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        requested_role: { type: string, enum: [reporter, technician] }
        status: { type: string, enum: [pending, approved, rejected, inactive] }
        registered_at: { type: string, format: date-time }
        approved_by: { oneOf: [{ $ref: '#/components/schemas/Reviewer' }, { type: 'null' }] }
        approved_at: { type: [string, 'null'], format: date-time }
        rejected_by: { oneOf: [{ $ref: '#/components/schemas/Reviewer' }, { type: 'null' }] }
        rejected_at: { type: [string, 'null'], format: date-time }
        rejection_reason: { type: [string, 'null'], maxLength: 1000 }
    AccountRequestSuccess:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/AccountRequest' }
...
    SuccessEnvelope:
      type: object
      required: [success, message, data, errors, code]
      properties:
        success: { const: true }
        message: { type: string }
        data: {}
        errors: { type: 'null' }
        code: { type: 'null' }
    ErrorEnvelope:
      type: object
      required: [success, message, data, errors, code]
      properties:
        success: { const: false }
        message: { type: string }
        data: { type: 'null' }
        errors:
          oneOf:
            - type: object
              additionalProperties:
                type: array
                items: { type: string }
            - type: 'null'
        code: { type: string }
  responses:
    ValidationError:
      description: Request validation failed with no state change
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    UnauthenticatedError:
      description: Credentials or active session are not accepted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    ForbiddenError:
      description: The authenticated account is not an administrator
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFoundError:
      description: The account request does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ConflictError:
      description: The account request is not pending and remains unchanged
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
security:
  - bearerAuth: []